requirements.txt
langchain
openai
python-dotenv
PyPDF2
streamlit